refactor(volume-splitter): extract service layer from route (WP1.7)
Move history/get/delete, range defaults/sanitization, and export orchestration
out of routes/volume_splitter.py into a new utils/volume_splitter_service.py.
The route handlers now parse/validate HTTP input, delegate to the service, and
shape responses with the same success_response()/error_response() envelopes,
status codes, and messages as before.

Behavior-preserving: no change to calculations, DB queries, response shapes, or
persistence. The service uses DatabaseHandler + get_logger() and does not import
from routes. The second volume-classification vocabulary used here is documented
in place and deliberately NOT consolidated with the canonical volume classes
(that is a later behavior-aware WP). The scan's noted silent-failure path is
preserved as-is (not changed).

Gate: import smoke OK (routes.volume_splitter + utils.volume_splitter_service).
Full pytest + E2E (volume-splitter.spec.ts, volume-progress.spec.ts) run on CI.
